apiCRUD/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
astroid==2.3.3
async-timeout==4.0.1
attrs==21.2.0
auditlog3==1.0.1
Babel==2.9.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bootstrap4==0.1.0
branca==0.4.2
Brlapi==0.7.0
CacheControl==0.12.11
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.8
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-braces==1.14.0
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-extra-keywords==1.1.2
django-js-asset==2.0.0
django-model-utils==4.0.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.16
duplicity==0.8.12.0
ebaysdk==2.1.5
entrypoints==0.3
fail2ban==0.11.1
fasteners==0.14.1
firebase-admin==5.2.0
flake8==5.0.4
freezegun==0.3.15
frozenlist==1.3.0
future==0.18.2
gevent==20.9.0
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
greenlet==0.4.17
grpcio==1.46.3
grpcio-status==1.46.3
gTTS==2.2.4
gunicorn==20.1.0
httplib2==0.20.4
idna==2.8
isodate==0.6.1
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libsass==0.18.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.5
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.7.0
monotonic==1.5
msgpack==1.0.4
multidict==5.2.0
netifaces==0.10.4
num2words==0.5.6
oauthlib==3.1.0
ofxparse==0.19
olefile==0.46
packaging==20.3
paramiko==2.6.0
passlib==1.7.3
pbr==5.9.0
pexpect==4.6.0
Pillow==9.0.1
polib==1.1.0
proto-plus==1.20.5
protobuf==3.6.1
psutil==5.6.7
psycopg2==2.8.6
pxpx==1.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.9.1
pycups==1.9.73
pydot==1.4.1
pyflakes==2.5.0
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyRFC3339==1.1
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-ldap==3.4.0
python-stdnum==1.13
pyttsx3==2.90
pytz==2019.3
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
ttystatus==0.38
typed-ast==1.4.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vobject==0.9.6.1
wadllib==1.3.3
Werkzeug==0.16.1
whitenoise==6.1.0
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
yarl==1.7.2
zeep==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
